[tests] Update invalid build command test.
